fix(sec+perf+ux): compile fixes, frontend noindex, cargo fmt
Verified:
- cargo check --features sqlite --lib --bin backend: clean
- cargo clippy --features sqlite --tests -- -D warnings: clean
- cargo fmt --all -- --check: clean
- cargo test config::tests (mask_db_url): 4 pass
- cargo test storage::local::tests (path traversal): 3 pass
- frontend tsc --noEmit: clean (only pre-existing baseUrl warning)

Changes in this commit:

Backend (sec+perf):
- booking_service.rs: fix confirm_inner closure capture (booking_id Uuid).
  Add hold_with_user + confirm_as_system; cancel/confirm take caller_user_id
  and verify booking.user_id == caller (BOLA defense).
- routes/bookings.rs: pass real uid to hold/cancel/confirm.
- routes/chat.rs: call assert_channel_access before list_messages,
  post_message, mark_read (BOLA defense on chat channels).
- service/chat_service.rs: add assert_channel_access helper.
- routes/zeroclaw.rs: list_exchanges requires AdminUser + RBAC.
- rbac/model.rs: add ADMIN_ZEROCLAW_READ permission.
- migrator seed: add admin:zeroclaw:read permission.
- ws/handler.rs: add check_ws_origin + parse_origin (CSWSH defense).
  Made check_ws_origin pub(crate) so audio_call can reuse it.
- audio_call/handler.rs: call check_ws_origin on every upgrade.
- storage/local.rs: component-walk path-traversal protection + tests.
- worker/db.rs: fix dead-code SKIP LOCKED (PERF-003). enqueue()
  calls wake.notify_waiters() (PERF-005). Add with_db constructor.
- config/mod.rs: mask_db_url() masks Postgres password in logs + config-show
  (STRUCT-013). Also mask OAuth + payment secrets. + unit tests.
- tests/api_smoke.rs: fix double-underscore env var names (STRUCT-014).
- payment_service.rs: use confirm_as_system for server-side confirm calls.
- service/review_service.rs: drop unnecessary i64-as-i64 casts
  (clippy::unnecessary_cast).
- audio_call/handler.rs: unwrap_or_else(|_| x) → unwrap_or(x)
  (clippy::unnecessary_lazy_evaluations).
- cargo fmt --all applied (was previously unformatted in places).

Frontend (UX):
- router.tsx: add <meta name=robots content=noindex,nofollow> for
  /admin, /account, /login routes (UIUX-017). Removed on public
  routes. Also skip GA4 page-view tracking on private routes.

Infra:
- .env.example: scrubbed real secrets.
- .gitignore: hardened (excludes *.db*, osm-index/, storage/, data/).
- terraform/deploy.sh: fix double-underscore env var names (SEC-003).
- Caddyfile: tighten CSP (no 'unsafe-eval', scoped wss://) (SEC-009).
- docker-compose.prod.yml: full rewrite (no host port for db/redis,
  no-new-privileges, cap_drop: ALL, resource limits, log caps) (SEC-010).
- .github/workflows/ci.yml: add fmt + test + audit + vitest + gitleaks +
  docker smoke build (STRUCT-015).
- migrator/m20260826_000001_audit_indexes.rs: new migration adding
  indexes for PERF-006, PERF-008-010, PERF-007 prep.
- migrator/src/migration.rs: register new migration.

Docs:
- AUDIT_REPORT.md: comprehensive audit summary.
- SECURITY.md: pre-deploy checklist + incident response.

+++ src/service/review_service.rs
@@ -289,7 +289,7 @@ impl ReviewService {
         let avg = if reviews.is_empty() {
             None
         } else {
-            let sum: i64 = reviews.iter().map(|r| r.rating as i64).sum();
+            let sum: i64 = reviews.iter().map(|r| r.rating).sum();
             Some(sum as f64 / reviews.len() as f64)
         };
 
@@ -335,13 +335,13 @@ pub fn review_to_dto(r: &review::Model) -> ReviewOut {
         brand_id: r.brand_id,
         author_name: r.author_name.clone(),
         author_phone: r.author_phone.clone(),
-        rating: r.rating as i64,
+        rating: r.rating,
         title: r.title.clone(),
         content: r.content.clone(),
         tags,
         photos,
         status: r.status.clone(),
-        helpful_count: r.helpful_count as i64,
+        helpful_count: r.helpful_count,
         reply: r.reply.clone(),
         replied_at: r.replied_at.clone(),
         created_at: r.created_at.clone(),
